+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `datacontract test` now supports testing HTTP APIs.
+
+
 ## [0.10.33] - 2025-07-29
 
 ### Added